+++ core/src/main/java/nl/nn/adapterframework/pipes/EtagHandlerPipe.java
@@ -24,10 +24,12 @@ import nl.nn.adapterframework.core.ParameterException;
 import nl.nn.adapterframework.core.PipeForward;
 import nl.nn.adapterframework.core.PipeRunException;
 import nl.nn.adapterframework.core.PipeRunResult;
-import nl.nn.adapterframework.http.IRestEtagCache;
-import nl.nn.adapterframework.http.RestEtagEhcache;
 import nl.nn.adapterframework.http.RestListenerUtils;
 import nl.nn.adapterframework.http.RestServiceDispatcher;
+import nl.nn.adapterframework.http.rest.ApiServiceDispatcher;
+import nl.nn.adapterframework.http.rest.ApiCacheManager;
+import nl.nn.adapterframework.http.rest.IApiCache;
+import nl.nn.adapterframework.http.rest.ApiEhcache;
 import nl.nn.adapterframework.parameters.Parameter;
 import nl.nn.adapterframework.parameters.ParameterList;
 import nl.nn.adapterframework.parameters.ParameterResolutionContext;
@@ -45,7 +47,7 @@ public class EtagHandlerPipe extends FixedForwardPipe {
 	List<String> actions = Arrays.asList("generate", "get", "set", "delete", "flush", "clear");
 	private String restPath = "/rest";
 	private String uriPattern = null;
-	private IRestEtagCache cache = null;
+	private IApiCache cache = null;
 
 	public void configure() throws ConfigurationException {
 		super.configure();
@@ -69,7 +71,7 @@ public class EtagHandlerPipe extends FixedForwardPipe {
 			throw new ConfigurationException(getLogPrefix(null)+"no uriPattern found!");
 		}
 
-		cache = RestServiceDispatcher.getCache();
+		cache = ApiCacheManager.getInstance();
 	}
 
 	public PipeRunResult doPipe(Object input, IPipeLineSession session) throws PipeRunException {
@@ -123,8 +125,8 @@ public class EtagHandlerPipe extends FixedForwardPipe {
 				returnCode = cache.remove(cacheKey);
 			}
 			else if(getAction().equalsIgnoreCase("flush")) {
-				if(cache instanceof RestEtagEhcache) {
-					((RestEtagEhcache) cache).flush();
+				if(cache instanceof ApiEhcache) {
+					((ApiEhcache) cache).flush();
 					returnCode = true;
 				}
 			}